// src/api/models/select-medication-record-hist-res.ts
/* tslint:disable */
/* eslint-disable */
/**
 * OpenAPI definition
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

/**
 *
 * @export
 * @interface SelectMedicationRecordHistRes
 */
export interface SelectMedicationRecordHistRes {
  /**
   *
   * @type {string}
   * @memberof SelectMedicationRecordHistRes
   */
  rowId?: string;
  /**
   *
   * @type {string}
   * @memberof SelectMedicationRecordHistRes
   */
  medicationId?: string;
  /**
   *
   * @type {string}
   * @memberof SelectMedicationRecordHistRes
   */
  medicationName?: string;
  /**
   *
   * @type {string}
   * @memberof SelectMedicationRecordHistRes
   */
  divisionCode?: SelectMedicationRecordHistResDivisionCodeEnum;
  /**
   *
   * @type {string}
   * @memberof SelectMedicationRecordHistRes
   */
  usageObject?: string;
  /**
   *
   * @type {string}
   * @memberof SelectMedicationRecordHistRes
   */
  prescriptionDate?: string | null;
  /**
   *
   * @type {number}
   * @memberof SelectMedicationRecordHistRes
   */
  prescriptionDays?: number;
  /**
   *
   * @type {string}
   * @memberof SelectMedicationRecordHistRes
   */
  unit?: string;
  /**
   *
   * @type {string}
   * @memberof SelectMedicationRecordHistRes
   */
  usageStatusCode?: SelectMedicationRecordHistResUsageStatusCodeEnum;
  /**
   *
   * @type {string}
   * @memberof SelectMedicationRecordHistRes
   */
  updatedDatetime?: string;
  /**
   *
   * @type {string}
   * @memberof SelectMedicationRecordHistRes
   */
  createdDatetime?: string;
  /**
   *
   * @type {string}
   * @memberof SelectMedicationRecordHistRes
   */
  createdBy?: string;
  /**
   *
   * @type {string}
   * @memberof SelectMedicationRecordHistRes
   */
  updatedBy?: string;
}

export const SelectMedicationRecordHistResDivisionCodeEnum = {
  Prescription: 'PRESCRIPTION',
  Otc: 'OTC',
} as const;

export type SelectMedicationRecordHistResDivisionCodeEnum =
  (typeof SelectMedicationRecordHistResDivisionCodeEnum)[keyof typeof SelectMedicationRecordHistResDivisionCodeEnum];
export const SelectMedicationRecordHistResUsageStatusCodeEnum = {
  Regular: 'REGULAR',
  AsNeeded: 'AS_NEEDED',
  Stopped: 'STOPPED',
} as const;

export type SelectMedicationRecordHistResUsageStatusCodeEnum =
  (typeof SelectMedicationRecordHistResUsageStatusCodeEnum)[keyof typeof SelectMedicationRecordHistResUsageStatusCodeEnum];